requirements.txt
aiogram==3.3.0
aiosqlite==0.20.0
asyncpg==0.29.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27